Moved Logic to lib
